add option to load house urls from a file
