+++ config.txt
@@ -1,0 +1,1 @@
+https://www.seabrookwa.com/vacation-rentals/beach-bluff
